+++ apps/air-discount-scheme/backend/src/app/modules/discount/dto/ExplicitFlight.dto.ts
@@ -7,6 +7,5 @@ export type ExplicitFlightLeg = {
 export type ExplicitFlight = {
   connectable: boolean
   id: string
-
   flightLegs: Array<ExplicitFlightLeg>
 }
