+++ apps/air-discount-scheme/backend/src/app/modules/flight/flight.service.ts
@@ -18,6 +18,7 @@ import {
 } from '../nationalRegistry'
 import { ExplicitCode } from '../discount/discount.model'
 import type { User as AuthUser } from '@island.is/auth-nest-tools'
+import { ExplicitFlightLeg } from '../discount/dto/ExplicitFlight.dto'
 
 export const ADS_POSTAL_CODES = {
   Reykhólahreppur: 380,
@@ -83,8 +84,8 @@ export class FlightService {
   }
 
   hasConnectingFlightPotentialFromFlightLegs(
-    firstFlight: FlightLeg,
-    secondFlight: FlightLeg,
+    firstFlight: FlightLeg | ExplicitFlightLeg,
+    secondFlight: FlightLeg | ExplicitFlightLeg,
   ): boolean {
     // If neither flight is connected to Reykjavik in any way
     // then it is not eligible
@@ -143,7 +144,7 @@ export class FlightService {
 
   async isFlightLegConnectingFlight(
     existingFlightId: string,
-    incomingLeg: FlightLeg,
+    incomingLeg: FlightLeg | ExplicitFlightLeg,
   ): Promise<boolean> {
     // Get the corresponding flight for the connection discount code
     const existingFlight = await this.flightModel.findOne({
