+++ libs/auth/scopes/src/lib/admin-portal.scope.ts
@@ -1,5 +1,6 @@
 export enum AdminPortalScope {
   airDiscountScheme = '@admin.island.is/ads',
+  explicitAirDiscountScheme = '@admin.island.is/ads:explicit',
   delegations = '@admin.island.is/delegations',
   regulationAdmin = '@admin.island.is/regulations',
   regulationAdminManage = '@admin.island.is/regulations:manage',
@@ -11,5 +12,4 @@ export enum AdminPortalScope {
   petitionsAdmin = '@admin.island.is/petitions',
   serviceDesk = '@admin.island.is/service-desk',
   signatureCollection = '@admin.island.is/signature-collection',
-  explicitAirDiscountScheme = '@admin.island.is/ads:explicit',
 }
